refactor: move backend files to app dir
