+++ src/backend/app/images/project_icon.png
