+++ src/backend/app/main.py
@@ -56,13 +56,6 @@ logging.basicConfig(
     stream=sys.stdout,
 )
 
-gunicorn_error_logger = logging.getLogger("gunicorn.error")
-gunicorn_logger = logging.getLogger("gunicorn")
-uvicorn_access_logger = logging.getLogger("uvicorn.access")
-uvicorn_access_logger.handlers = gunicorn_error_logger.handlers
-
-fastapi_logger.handlers = gunicorn_error_logger.handlers
-
 logger = logging.getLogger(__name__)
